Refactor CassandraPersistence to split out Task and Shard Persistence
