+++ common/persistence/cassandra/cassandraShardPersistence.go
@@ -1,0 +1,244 @@
+// Copyright (c) 2019 Uber Technologies, Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package cassandra
+
+import (
+	"context"
+	"fmt"
+	"strings"
+	"time"
+
+	"github.com/gocql/gocql"
+
+	workflow "github.com/uber/cadence/.gen/go/shared"
+	"github.com/uber/cadence/common/log"
+	p "github.com/uber/cadence/common/persistence"
+	"github.com/uber/cadence/common/persistence/nosql/nosqlplugin/cassandra"
+	"github.com/uber/cadence/common/service/config"
+)
+
+type (
+	// Implements ShardManager
+	cassandraShardPersistence struct {
+		cassandraStore
+		shardID            int
+		currentClusterName string
+	}
+)
+
+var _ p.ShardStore = (*cassandraShardPersistence)(nil)
+
+// newShardPersistence is used to create an instance of ShardManager implementation
+func newShardPersistence(cfg config.Cassandra, clusterName string, logger log.Logger) (p.ShardStore, error) {
+	cluster := cassandra.NewCassandraCluster(cfg)
+	cluster.ProtoVersion = cassandraProtoVersion
+	cluster.Consistency = gocql.LocalQuorum
+	cluster.SerialConsistency = gocql.LocalSerial
+	cluster.Timeout = defaultSessionTimeout
+
+	session, err := cluster.CreateSession()
+	if err != nil {
+		return nil, err
+	}
+
+	return &cassandraShardPersistence{
+		cassandraStore:     cassandraStore{session: session, logger: logger},
+		shardID:            -1,
+		currentClusterName: clusterName,
+	}, nil
+}
+
+// NewShardPersistenceFromSession is used to create an instance of ShardManager implementation
+// It is being used by some admin toolings
+func NewShardPersistenceFromSession(session *gocql.Session, clusterName string, logger log.Logger) p.ShardStore {
+	return &cassandraShardPersistence{
+		cassandraStore:     cassandraStore{session: session, logger: logger},
+		shardID:            -1,
+		currentClusterName: clusterName,
+	}
+}
+
+func (d *cassandraShardPersistence) CreateShard(
+	_ context.Context,
+	request *p.InternalCreateShardRequest,
+) error {
+	cqlNowTimestamp := p.UnixNanoToDBTimestamp(time.Now().UnixNano())
+	shardInfo := request.ShardInfo
+	markerData, markerEncoding := p.FromDataBlob(shardInfo.PendingFailoverMarkers)
+	transferPQS, transferPQSEncoding := p.FromDataBlob(shardInfo.TransferProcessingQueueStates)
+	timerPQS, timerPQSEncoding := p.FromDataBlob(shardInfo.TimerProcessingQueueStates)
+	query := d.session.Query(templateCreateShardQuery,
+		shardInfo.ShardID,
+		rowTypeShard,
+		rowTypeShardDomainID,
+		rowTypeShardWorkflowID,
+		rowTypeShardRunID,
+		defaultVisibilityTimestamp,
+		rowTypeShardTaskID,
+		shardInfo.ShardID,
+		shardInfo.Owner,
+		shardInfo.RangeID,
+		shardInfo.StolenSinceRenew,
+		cqlNowTimestamp,
+		shardInfo.ReplicationAckLevel,
+		shardInfo.TransferAckLevel,
+		shardInfo.TimerAckLevel,
+		shardInfo.ClusterTransferAckLevel,
+		shardInfo.ClusterTimerAckLevel,
+		transferPQS,
+		transferPQSEncoding,
+		timerPQS,
+		timerPQSEncoding,
+		shardInfo.DomainNotificationVersion,
+		shardInfo.ClusterReplicationLevel,
+		shardInfo.ReplicationDLQAckLevel,
+		markerData,
+		markerEncoding,
+		shardInfo.RangeID)
+
+	previous := make(map[string]interface{})
+	applied, err := query.MapScanCAS(previous)
+	if err != nil {
+		if isThrottlingError(err) {
+			return &workflow.ServiceBusyError{
+				Message: fmt.Sprintf("CreateShard operation failed. Error: %v", err),
+			}
+		}
+		return &workflow.InternalServiceError{
+			Message: fmt.Sprintf("CreateShard operation failed. Error: %v", err),
+		}
+	}
+
+	if !applied {
+		shard := previous["shard"].(map[string]interface{})
+		return &p.ShardAlreadyExistError{
+			Msg: fmt.Sprintf("Shard already exists in executions table.  ShardId: %v, RangeId: %v",
+				shard["shard_id"], shard["range_id"]),
+		}
+	}
+
+	return nil
+}
+
+func (d *cassandraShardPersistence) GetShard(
+	_ context.Context,
+	request *p.InternalGetShardRequest,
+) (*p.InternalGetShardResponse, error) {
+	shardID := request.ShardID
+	query := d.session.Query(templateGetShardQuery,
+		shardID,
+		rowTypeShard,
+		rowTypeShardDomainID,
+		rowTypeShardWorkflowID,
+		rowTypeShardRunID,
+		defaultVisibilityTimestamp,
+		rowTypeShardTaskID)
+
+	result := make(map[string]interface{})
+	if err := query.MapScan(result); err != nil {
+		if err == gocql.ErrNotFound {
+			return nil, &workflow.EntityNotExistsError{
+				Message: fmt.Sprintf("Shard not found.  ShardId: %v", shardID),
+			}
+		} else if isThrottlingError(err) {
+			return nil, &workflow.ServiceBusyError{
+				Message: fmt.Sprintf("GetShard operation failed. Error: %v", err),
+			}
+		}
+
+		return nil, &workflow.InternalServiceError{
+			Message: fmt.Sprintf("GetShard operation failed. Error: %v", err),
+		}
+	}
+
+	info := createShardInfo(d.currentClusterName, result["shard"].(map[string]interface{}))
+
+	return &p.InternalGetShardResponse{ShardInfo: info}, nil
+}
+
+func (d *cassandraShardPersistence) UpdateShard(
+	_ context.Context,
+	request *p.InternalUpdateShardRequest,
+) error {
+	cqlNowTimestamp := p.UnixNanoToDBTimestamp(time.Now().UnixNano())
+	shardInfo := request.ShardInfo
+	markerData, markerEncoding := p.FromDataBlob(shardInfo.PendingFailoverMarkers)
+	transferPQS, transferPQSEncoding := p.FromDataBlob(shardInfo.TransferProcessingQueueStates)
+	timerPQS, timerPQSEncoding := p.FromDataBlob(shardInfo.TimerProcessingQueueStates)
+
+	query := d.session.Query(templateUpdateShardQuery,
+		shardInfo.ShardID,
+		shardInfo.Owner,
+		shardInfo.RangeID,
+		shardInfo.StolenSinceRenew,
+		cqlNowTimestamp,
+		shardInfo.ReplicationAckLevel,
+		shardInfo.TransferAckLevel,
+		shardInfo.TimerAckLevel,
+		shardInfo.ClusterTransferAckLevel,
+		shardInfo.ClusterTimerAckLevel,
+		transferPQS,
+		transferPQSEncoding,
+		timerPQS,
+		timerPQSEncoding,
+		shardInfo.DomainNotificationVersion,
+		shardInfo.ClusterReplicationLevel,
+		shardInfo.ReplicationDLQAckLevel,
+		markerData,
+		markerEncoding,
+		shardInfo.RangeID,
+		shardInfo.ShardID,
+		rowTypeShard,
+		rowTypeShardDomainID,
+		rowTypeShardWorkflowID,
+		rowTypeShardRunID,
+		defaultVisibilityTimestamp,
+		rowTypeShardTaskID,
+		request.PreviousRangeID)
+
+	previous := make(map[string]interface{})
+	applied, err := query.MapScanCAS(previous)
+	if err != nil {
+		if isThrottlingError(err) {
+			return &workflow.ServiceBusyError{
+				Message: fmt.Sprintf("UpdateShard operation failed. Error: %v", err),
+			}
+		}
+		return &workflow.InternalServiceError{
+			Message: fmt.Sprintf("UpdateShard operation failed. Error: %v", err),
+		}
+	}
+
+	if !applied {
+		var columns []string
+		for k, v := range previous {
+			columns = append(columns, fmt.Sprintf("%s=%v", k, v))
+		}
+
+		return &p.ShardOwnershipLostError{
+			ShardID: d.shardID,
+			Msg: fmt.Sprintf("Failed to update shard.  previous_range_id: %v, columns: (%v)",
+				request.PreviousRangeID, strings.Join(columns, ",")),
+		}
+	}
+
+	return nil
+}
